backend/requirements.txt
flask
flask-cors
pymongo
dnspython
pyjwt
python-dotenv
numpy
